Solved problem 4

+++ ProjectEuler/LargestPrimeFactor.java
@@ -1,5 +1,3 @@
-import java.math.BigInteger;
-
 /**
  * Problem: https://projecteuler.net/problem=3
  * The prime factors of 13195 are 5, 7, 13 and 29.
@@ -23,6 +21,6 @@ public class LargestPrimeFactor {
             }
         }
         while (n>=p*p);
-        System.out.println(n);
+        System.out.println((int)n);
     }
 }
